rename range->range_map
