+++ train-sim/libtsim/src/lib.rs
@@ -2,7 +2,7 @@
 
 #![warn(missing_docs)]
 
-pub use crate::range::RangeMap;
+pub use crate::range_map::RangeMap;
 
 mod integration;
-mod range;
+mod range_map;
