+++ train-sim/libtsim/src/range_map.rs
